requirements.txt

dagster-dbt
dagster-airbyte
dagster-embedded-elt
dagster-duckdb
dagster-duckdb-pandas
dagster-duckdb-polars
duckdb
sqlescapy
pandas
polars
pyarrow
geopandas
scikit-learn
fancyimpute
shapely
urllib3
minio
psycopg2-binary
mysql-connector-python
urllib3
minio